before big changes

+++ Models/Cart.cs
@@ -4,7 +4,6 @@ public class Cart
 {
     public int Id { get; set; }
 
-    //  user_id та session_id nullable для  кошика гостей 
     public int? UserId { get; set; }
     public User? User { get; set; }
 
@@ -13,6 +12,6 @@ public class Cart
     public int ProductId { get; set; }
     public Product Product { get; set; } = null!;
 
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public DateTime AddedAt { get; set; }
 }
